Zombie Game Update
Added various files:
Created a class for the smart agent
This class hold information and pointers to the behaviours. However these cannot be updated from this class hence they are for the moment updated in the plugin.
Blackboard was added and various other data into it.

More work needs to be done